requirements.txt
streamlit
transformers[sentencepiece]
torch
torchvision
torchaudio
Pillow
easyocr
numpy
PyMuPDF